class activity to create methods to return square of numbers and 2D array

+++ java/JAVA/src/ActivityOne.java
@@ -1,5 +1,5 @@
 public class ActivityOne {
-    static double sqr(double a){
+    static int sqr(int a){
         return a * a;
     }
